+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.1.0] - 2024-02-14
 ### Changed
 - Added NullIs option to the BooleanToBooleanConverter.
+- Added NullIs option to the BooleanToVisibilityConverter.
 ### Supported .Net Versions
 - .Net 6
 - .Net 7